Commit #0006 Huge recreation
This is a huge update, it has more structure for future
expansion. Implementation of all functions has been greatly
improved. Still few issues with multiple windows, but will address
them later. The save/load function interaction between the user and
the program needs to be remade slightly, and a good way of keeping
multiple buffers needs to be implemented.